game\server: Do not remove lag record too early
See https://github.com/ValveSoftware/source-sdk-2013/pull/884

+++ game/server/player_lagcompensation.cpp
@@ -251,7 +251,8 @@ void CLagCompensationManager::FrameUpdatePostEntityThink()
 	VPROF_BUDGET( "FrameUpdatePostEntityThink", "CLagCompensationManager" );
 
 	// remove all records before that time:
-	int flDeadtime = gpGlobals->curtime - sv_maxunlag.GetFloat();
+	// dimhotepus: int -> float to not remove lag record too early.
+	float flDeadtime = gpGlobals->curtime - sv_maxunlag.GetFloat();
 
 	// Iterate all active players
 	for ( int i = 1; i <= gpGlobals->maxClients; i++ )
